restructered signal module and added new opt module
