+++ modopt/opt/cost.py
