+++ modopt/opt/linear.py
